CrystalStructurePrediction

Documentation for CrystalStructurePrediction.

CrystalStructurePrediction.Ion

CrystalStructurePrediction.Lattice

CrystalStructurePrediction.build_ion_list

CrystalStructurePrediction.build_quadratic_problem

CrystalStructurePrediction.real_space_sum

CrystalStructurePrediction.summation

CrystalStructurePrediction.Ion — Type

Ion{D, T}
Ion(species::Symbol, charge::Int, radii::Float64, frac_pos)

An ion is a species with a charge, a radius, and a fractional position. D is the dimension of the space, and T is the type of the coordinates.

Fields

species::Symbol: The species of the ion.

charge::Int: The charge of the ion.

radii::Float64: The radius of the ion.

frac_pos::SVector{D, T}: The fractional position of the ion.

Examples

To define an ion with species "O", charge -2, radius 1.35, and fractional position [0.25, 0.25, 0.25], run

julia> Ion(:O, -2, 1.35, [0.25, 0.25, 0.25])
Ion{3, Float64}(:O, -2, 1.35, [0.25, 0.25, 0.25])

CrystalStructurePrediction.Lattice — Type

Lattice{D, T, L}
Lattice(vectors::AbstractMatrix{T}, pbc::NTuple{D, Bool}) where {D, T}

A lattice is a set of vectors that define the unit cell of a crystal.

Fields

vectors::SMatrix{D, D, T, L}: The vectors that define the unit cell of a crystal.

pbc::NTuple{D, Bool}: The periodic boundary conditions.

CrystalStructurePrediction.build_ion_list — Method

return a list of ions in the format of: [ion(t1, p1), ion(t1, p2), ..., ion(t1,pn), ion(t2, p1), ..., ion(tm, pn)]

CrystalStructurePrediction.build_quadratic_problem — Method

build_quadratic_problem(
    grid_size::NTuple{N, Int},
    population_list::AbstractVector{Int},
    interaction_matrix::AbstractMatrix{T},
    optimizer
    ) where {N, T<:Real}

Build a quadratic problem for crystal structure prediction.

Arguments

grid_size::NTuple{N, Int}: The size of the grid.

population_list::AbstractVector{Int}: The number of atoms of each species.

interaction_matrix::AbstractMatrix{T}: The interaction matrix.

optimizer: The optimizer.

CrystalStructurePrediction.real_space_sum — Method

real_space_sum(depth, frac_pos_a, frac_pos_b, lattice, alpha)

Compute the real space part of Ewald summation.

Argument: lattice::AbstractMatrix: in the form [a, b, c], where a, b, c are column vectors.

CrystalStructurePrediction.summation — Method

summation(depth, interaction)

Compute periodic summation of interaction to given depth.

Argument: depth::AbstractVector{Int}: summation depth on the dimensions. interaction<:Function: the interaction energy, given the shift as a list of integer.