CrystalStructurePrediction

Documentation for CrystalStructurePrediction.

CrystalStructurePrediction.Ion

CrystalStructurePrediction.IonType

CrystalStructurePrediction.Lattice

CrystalStructurePrediction.interaction_energy

CrystalStructurePrediction.ions_on_grid

CrystalStructurePrediction.optimize_linear

CrystalStructurePrediction.optimize_quadratic

CrystalStructurePrediction.periodic_sum

CrystalStructurePrediction.Ion — Type

Ion{D, T}
Ion(type::IonType{T}, frac_pos)

An ion is a type with a fractional position. D is the dimension of the space, and T is the type of the coordinates.

Fields

type::IonType{T}: The type of the ion.

frac_pos::SVector{D, T}: The fractional position of the ion.

CrystalStructurePrediction.IonType — Type

IonType{T}
IonType(species::Symbol, charge::Int, radii::T)

An ion type is a species with a charge and a radius. T is the type of the coordinates.

Fields

species::Symbol: The species of the ion.

charge::Int: The charge of the ion.

radii::T: The radius of the ion.

Examples

To define an ion type with species "O", charge -2, radius 1.35, run

julia> IonType(:O, -2, 1.35)
IonType{Float64}(:O, -2, 1.35)

CrystalStructurePrediction.Lattice — Type

Lattice{D, T, L}
Lattice(vectors::AbstractMatrix{T}, pbc::NTuple{D, Bool}) where {D, T}

A lattice is a set of vectors that define the unit cell of a crystal.

Fields

vectors::SMatrix{D, D, T, L}: The vectors that define the unit cell of a crystal.

pbc::NTuple{D, Bool}: The periodic boundary conditions.

CrystalStructurePrediction.interaction_energy — Method

interaction_energy(ion_a, ion_b, lattice, alpha, real_depth, reciprocal_depth, buckingham_depth, buckingham_threshold, buckingham_penalty, radii_threshold, radii_penalty)

Compute the interaction energy between two ions, which is a sum of the real space Ewald sum, the reciprocal space Ewald sum, the Buckingham potential, and the radii penalty.

Arguments

ion_a::Ion{D, T}: the first ion.

ion_b::Ion{D, T}: the second ion.

lattice::Lattice{D, T}: the lattice.

alpha::T: the Ewald parameter.

real_depth::NTuple{D, Int}: summation depth on the dimensions for the real space Ewald sum.

reciprocal_depth::NTuple{D, Int}: summation depth on the dimensions for the reciprocal space Ewald sum.

buckingham_depth::NTuple{D, Int}: summation depth on the dimensions for the Buckingham potential.

Keyword arguments

buckingham_threshold::T=0.75: the threshold for the Buckingham potential.

buckingham_penalty::T=3e2: the penalty for the Buckingham potential.

radii_threshold::T=0.0: the threshold for the radii penalty, defined as the ratio of the distance between the two ions to the sum of their radii.

radii_penalty::T=0.0: the penalty for the radii penalty.

References

a great short note: https://www.cs.cornell.edu/courses/cs428/2006fa/Ewald%20Sum.pdf

CrystalStructurePrediction.ions_on_grid — Method

ions_on_grid(grid_size::NTuple{N, Int}, type_list::AbstractVector{IonType{T}}) where {N, T}

Create a list of ions on a grid.

Arguments

grid_size::NTuple{N, Int}: the size of the grid.

type_list::AbstractVector{IonType{T}}: the list of ion types.

Returns

A vector of ions: [ion(t1, p1), ion(t1, p2), ..., ion(t1,pn), ion(t2, p1), ..., ion(tm, pn)].

where t1, ..., tm are the types of the ions and p1, ..., pn are the fractional positions of the ions on the grid.

CrystalStructurePrediction.optimize_linear — Method

optimize_linear(interaction, ions, populations, lattice; optimizer = SCIP.Optimizer, optimizer_options = Dict(), proximal_threshold = 0.75)

Build a linear problem for crystal structure prediction. Suited for solvers not supporting quadratic constraints.

Arguments

interaction::Function: The interaction function.

ions::AbstractVector{Ion{D, T}}: The ions.

populations::Dict{IonType{T}, Int}: The populations of the ions.

lattice::Lattice: The lattice.

Keyword arguments

optimizer: The optimizer.

optimizer_options: The options for the optimizer, e.g. optimizer_options = Dict("NodefileSave" => 1) for Gurobi.

proximal_threshold: The threshold for the proximal pairs.

CrystalStructurePrediction.optimize_quadratic — Method

optimize_quadratic(interaction, ions, populations, lattice; optimizer = SCIP.Optimizer, optimizer_options = Dict(), proximal_threshold = 0.75)

Build a quadratic problem for crystal structure prediction.

Arguments

interaction::Function: The interaction function.

ions::AbstractVector{Ion{D, T}}: The ions.

populations::Dict{IonType{T}, Int}: The populations of the ions.

lattice::Lattice: The lattice.

Keyword arguments

optimizer: The optimizer.

optimizer_options: The options for the optimizer, e.g. optimizer_options = Dict("NodefileSave" => 1) for Gurobi.

proximal_threshold: The threshold for the proximal pairs.

CrystalStructurePrediction.periodic_sum — Method

periodic_sum(interaction, depth)

Compute periodic summation of interaction to given depth.

Arguments

interaction<:Function: the interaction energy, given the shift as a list of integers.

depth::NTuple{N, Int}: summation depth on the dimensions.